pyproject.toml
[project]
name = "da-forecast"
version = "0.1.0"
description = "Day-ahead electricity price forecasting for European bidding zones"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "xgboost>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "entsoe-py>=0.6",
    "networkx>=3.1",
    "holidays>=0.40",
    "jupyter>=1.0",
    "ipykernel>=6.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
